Refactor remaining circular imports;
